requirements.txt
keras
pydot
tensorflow
pillow
scikit-image